Fixed the part regarding the cylindrical dish
Before the key was not working if a cylindrical dish was not being used.
The space for the connector was not removed correctly from the key. Now
if a cylindrical shape is not being used, a cube is used to remove the
connector part from the top of the key.